+++ src/main/java/com/cdkj/gchf/bo/impl/UserBOImpl.java
@@ -213,7 +213,7 @@ public class UserBOImpl extends PaginableBOImpl<User> implements IUserBO {
             Employ employCondition = new Employ();
             employCondition.setProjectCodeList(projectCodeList);
             employCondition.setStatus(EEmployStatus.Not_Leave.getCode());
-            data.setEmployCount(employBO.getTotalCount(employCondition));
+            data.setEmployCount((int) employBO.getTotalCount(employCondition));
         }
 
         // 业主用户数据
